requirements.txt
# ADHD项目依赖库

# 核心深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 数据处理和科学计算
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# 数据读取
pyreadstat>=1.1.0  # 用于读取STATA文件

# 可视化
matplotlib>=3.3.0
seaborn>=0.11.0

# 配置文件处理
PyYAML>=5.4.0

# 统计分析
scipy>=1.7.0

# 进度条
tqdm>=4.62.0

# Jupyter支持（可选）
jupyter>=1.0.0
ipykernel>=6.0.0